+++ CHANGES.md
@@ -51,7 +51,8 @@
 - Renamed `data` to `fy` in `RegPull.evaluate` and `RegAsProxyPull.evaluate`
 - Made `get_layer` in `FullyConnected` private
 - Made `get_dense` and `load_embeds` in `CatEmbHead` private
-Made `build_layers` in 'ClassRegMulti` private
+- Made `build_layers` in 'ClassRegMulti` private
+- Made parse methods and `build_opt` in `ModelBuilder` private
 
 
 ## Depreciations